increased day length (24h)

+++ src/texteff.cpp
@@ -37,7 +37,7 @@ struct TextEffect : public ViewportSign {
 static SmallVector<struct TextEffect, 32> _text_effects; ///< Text effects are stored there
 
 /* Text Effects */
-TextEffectID AddTextEffect(StringID msg, int center, int y, uint8 duration, TextEffectMode mode)
+TextEffectID AddTextEffect(StringID msg, int center, int y, int duration, TextEffectMode mode)
 {
 	if (_game_mode == GM_MENU) return INVALID_TE_ID;
